requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
dictdiffer==0.8.1
Django==3.2.5
djangorestframework==3.12.4
future==0.18.2
idna==2.10
mock==4.0.3
pytz==2021.1
requests==2.25.1
serializers==0.2.4
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.6
views==0.3
